refactor: adding plotting.py
