+++ openadapt/app/dashboard/api/recordings.py
@@ -7,7 +7,8 @@ from openadapt.app.cards import is_recording, quick_record, stop_record
 from openadapt.db import crud
 from openadapt.events import get_events
 from openadapt.models import Recording
-from openadapt.utils import display_event, image2utf8, row2dict
+from openadapt.utils import image2utf8, row2dict
+from openadapt.plotting import display_event
 
 
 class RecordingsAPI:
