+++ openadapt/productivity.py
@@ -21,9 +21,9 @@ from loguru import logger
 from openadapt.db.crud import get_latest_recording, get_window_events
 from openadapt.events import get_events
 from openadapt.models import ActionEvent, WindowEvent
+from openadapt.plotting import display_event,
 from openadapt.utils import (
     configure_logging,
-    display_event,
     image2utf8,
     row2dict,
     rows2dicts,
